KAN-48 feat: add ClientRooltLayout
