+++ src/app/layout.tsx
@@ -3,6 +3,7 @@ import { Inter } from 'next/font/google'
 import './globals.css'
 import { StoreProvider } from '@/shared/store/providers'
 import { Header } from '@/widgets/Header'
+import ClientRootLayout from './ClientRootLayout'
 
 const inter = Inter({
    variable: '--font-inter',
@@ -50,7 +51,7 @@ export default function RootLayout({
          <html lang="en">
             <body className={`${inter.variable}`}>
                <Header notificationCount={4} />
-               {children}
+               <ClientRootLayout>{children}</ClientRootLayout>
             </body>
          </html>
       </StoreProvider>
